seteo y resta de puntos de beneficio para luego usarlos en el microservicio de facturacion

+++ micro_Cliente/src/main/java/com/example/micro_cliente/entities/ClienteEntity.java
@@ -35,5 +35,5 @@ public class ClienteEntity {
     @Column
     @Min(0)
     @Max(5000)
-    private Long puntosBeneficio;
+    private Double puntosBeneficio;
 }
