+++ micro_Cliente/src/main/java/com/example/micro_cliente/models/Cliente.java
@@ -16,5 +16,5 @@ public class Cliente {
     private String domicilio;
     private String telefono;
     private String email;
-    private Long puntosBeneficio;
+    private Double puntosBeneficio;
 }
